Reversed Validator/Node dependency
Why: In order to widen the context of Validator and Node, Node now passes itself as the candidate to any validators. This avoids Node having to understand what
Validator requires in order to do its job. Validator still knows nothing about node.

This change addresses the need by:

Validator now accepts a key option in its constructor which it uses to assert the presence of the value it is supposed to validate from the payload passed into it.
It also accepts any object that responds to #[] as a candidate for validation. Node now responds to [] by delegating this method to its @content attribute.

There were several other code improvements to the classes.

Something that really shone out during this refactoring was the importance of reducing dependencies by wrapping alien objects and their methods, regardless of
how trivial that may be. The primary driver was that in Node @content is usually a Hash procured from a JSON payload. The keys in a JSON payload are usually strings
(indeed it is worth noting that Node expects this). However, the keys that are passed in are usually accepted as symbols, such as :key. I had considerable anguish
during the refactoring because I kept forgetting this distinction and tried to access @content using the key as a symbol. I had felt that I ought not to rely on
Hash's interface but I thought "Well, it's very stable so it should be OK". Wrong! Always wrap objects of a different type. Always. Then you can specify the context
in which they will be used in great detail.