Added ContentWrapper
Why: Content will almost always be sourced from JSON, but the resulting Hash's keys are usually strings. In order
to avoid having to remember to use Hash#[key.to_s] in the right place, and all the bugs it could produce, I have
created a class that the library can use to wrap content. It responds to [], except that symbol arguments are converted
to keys before the call is forwarded to the wrapped object