Country Payload now validating
Why: Country payload did not validate properly because of a bug in SchemaGenerator.
EAS-15 #comment SchemaGenerator did not add the correct key to the schema when it was retrieved from the library. This has been corrected.
A side effect of this specific bug was discovering that ArrayValidator did not handle nil values correctly - it threw a type error instead of returning false
EAS-16 #done ArrayValidator refactored to handle TypeError as well as ArgumentError
As an additional bonus I have also refactored SchemaLibrary to return a copy of any schemas requested rather than the original.