Removed cotains_schema and passed responsibility to has_attribute
Why:
The contains_schema method explicitly set the key on the retrieved schema procured from ValidatorFactory. This was because a schema, when retrieved from the library had no key set. This was a developer-induced dependency: I knew that the schemas in the schema library had no key, so I passed that knowledge on to SchemaGenerator.

This change addresses the need by: Removing the contains_schema method from SchemaGenerator and stipulating that type: :schema and name: <name> must be added to the opts hash. ValidatoryFactory now handles the interface with SchemaLibrary, and SchemaLibrary now handles setting the key on the schema. This is not yet perfect, but moves the dependency for key a lot closer to the schema instance that is to be used.