Added type specific has_* methods to SchemaGenerator
Why: The initial interface of SchemaGenerator used the #has_attribute method as a way of adding attributes to a schema. The problem with this is that this leaves clients responsible for defining the type by passing a :type argument in the opts hash. So the client not only has to know that there is a type of value in its schem, but also that there is a specific keyword to tell the library that its there. Given the JSON_Schema is a very stable language it makes more sense to add specific methods to the JSONSchema interface in order to hide the idea of a keyword from the client.

This fix addresses the need by: Adding keyword specific addition methods to the SchemaGenerator interface. The has_attribute method has not been removed for reasons of backwards compatibility but will probably be deprecated in the future or at least marked as private