Implemented Mechanism for Error Messages
Validators must now set an error message when they find an invalid value

This commit makes some large changes to the internals of the validators, partly due to a change in the validation process, and partly due to the unearthing of some bugs whilst implementing the validation process.

The main point are as follows:

`Validator` now interprets the results of the validation operation based on the content of the `errors` array. If there is anything in it, the node is assumed to be invalid. This is similar to the approach taken by `ActiveRecord`.

`ContentWrapper` is now a module which is included in `Node`, rather than an individual class. I had always been slightly nervous of wrapping JSON content in `Node` with `ContentWrapper` and I was right - the effect was that content were being repeatedly wrapped, creating inaccessible values and a false representation of the schema's structure. By including the functionality of `ContentWrapper` in `Node` I removed this additional complexity by essentially making `Node` the wrapper, which it always was, really.

This branch is not finished yet. Part of the reason for this commit is to allow me to continue editing on a different computer.