ValidatorSet Implemented
Added ValidatorSet and established that it works correctly.
Installed ValidatorSet in Node and established that it satisfies
correctly replaces the validation functionality of the validator set,
with the exception of :strict. This is to be implemented in the next
commit by upgrading ObjectValidator

Notes:

This was an excellent demonstration of the costs introduced by a
misplaced addition to an inheritance chain. Node, as noted elsewhere,
was not really a Validator. It was an aggregator of Validators. However,
by inheriting from Validator I created a spaghetti mess which was
difficult to extract Node from since it relied so heavily on Validator's
validation system.

I had suspected this was a problem early on, but did not deal with it as
it was not causing any issues at that point. In retrospect I should
have. There were three early signs that I put on the backburner. One was
the realisation that the validators array was probably too general an
abstraction and that something more concrete may be necessary, the
second was a general sense of lack of clarity with the role of Node and
the third was the amount of validator methods that had to be overriden
to make Node work. I think of all of them, the first two were the most
worthy of attention since they occured prior to the code being written
but the third should have been seen as an absolute red flag and seen as
a show stopper worthy of delaying publication.