Validators now only require the candidate value
Why: Thew original implementation of validator accepted a hash as its
candidate, the logic being that the Validator would extract the relevant
value required from its key. This created a variety of issues since it
made added the responsibility of extracting the value to the Validator
and in doing so added additional requirements to the context which the
validator had to be used in.

This fix addresses the need by: Removing the requirement for the key and
only asking Validator to validate the value that it has been passed.
Responsibility for extracting the value now sites with the client.