Implemented ValidationStep and Tested
ValidationStep provides a method of chaining together verification steps
for validating a value. By creating this chain we can create reusable
validators which could eventually supercede the Validator class.
However, for the time being it is probably still best to use
ValidationSteps within the Validator class. They could certainly be used
to get rid of the inheritance hierarchy of validator and use composition
instead.