Validation is now Stateless
Why: State is a pain, especially if it changes. You have objects flying
around which can be in an inconsistent state. This problem has already
been seen in early versions of the gem where a reset! method was added
to clear error messages.

The fix addresses the need by:
Recursively building up the error array from the lowest points in the
verification tree.