Added use_schema to SchemaLibrary interface
In my rush to get all the tests to pass, I had manually wrapped the
schema with a `Validator` instance that is used in "As a user I want to
reuse a saved schema" which undermines the point of having a DSL.

This fix adds the `use_schema` method to `SchemaLibrary` so that a
wrapper instance is returned. This is the correct way to retrieve a
schema for reuse outwith another schema.